requirements.txt
fastapi==0.68.2
pydantic==1.8.2
uvicorn==0.15.0
python-dateutil==2.8.2
pytest==6.2.5
pytest-cov==3.0.0
requests==2.26.0